Move caps.Type to a dedicated file
